backend/requirements.txt
# Django
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.14,<4.0

# JWT Authentication
djangorestframework-simplejwt>=5.3,<6.0

# CORS Headers
django-cors-headers>=4.3,<5.0

# PostgreSQL adapter
psycopg2-binary>=2.9,<3.0

# Environment variables
python-dotenv>=1.0,<2.0
